+++ app/src/main/AndroidManifest.xml
@@ -335,11 +335,6 @@
                 <data android:scheme="yuedu" />
             </intent-filter>
         </activity>
-        <!-- 导入书源 -->
-        <activity
-            android:name=".ui.association.ImportBookSourceActivity"
-            android:configChanges="locale|keyboardHidden|orientation|screenSize|smallestScreenSize|screenLayout"
-            android:theme="@style/AppTheme.Transparent" />
         <!-- 导入订阅源 -->
         <activity
             android:name=".ui.association.ImportRssSourceActivity"
