+++ app/src/main/java/io/legado/app/ui/association/ImportBookSourceDialog.kt
@@ -2,14 +2,14 @@ package io.legado.app.ui.association
 
 import android.annotation.SuppressLint
 import android.content.Context
-import android.content.DialogInterface
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.widget.Toolbar
-import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import io.legado.app.R
 import io.legado.app.base.BaseDialogFragment
@@ -23,13 +23,11 @@ import io.legado.app.databinding.DialogEditTextBinding
 import io.legado.app.databinding.DialogRecyclerViewBinding
 import io.legado.app.databinding.ItemSourceImportBinding
 import io.legado.app.help.AppConfig
+import io.legado.app.help.IntentDataHelp
 import io.legado.app.lib.dialogs.alert
 import io.legado.app.ui.widget.dialog.WaitDialog
-import io.legado.app.utils.dp
-import io.legado.app.utils.putPrefBoolean
-import io.legado.app.utils.splitNotBlank
+import io.legado.app.utils.*
 import io.legado.app.utils.viewbindingdelegate.viewBinding
-import io.legado.app.utils.visible
 
 
 /**
@@ -37,10 +35,21 @@ import io.legado.app.utils.visible
  */
 class ImportBookSourceDialog : BaseDialogFragment(), Toolbar.OnMenuItemClickListener {
 
-    private val binding by viewBinding(DialogRecyclerViewBinding::bind)
+    companion object {
 
-    val viewModel: ImportBookSourceViewModel by activityViewModels()
-    lateinit var adapter: SourcesAdapter
+        fun start(fragmentManager: FragmentManager, source: String) {
+            ImportBookSourceDialog().apply {
+                arguments = Bundle().apply {
+                    putString("source", source)
+                }
+            }.show(fragmentManager, "importBookSource")
+        }
+
+    }
+
+    private val binding by viewBinding(DialogRecyclerViewBinding::bind)
+    private val viewModel by viewModels<ImportBookSourceViewModel>()
+    private lateinit var adapter: SourcesAdapter
 
     override fun onStart() {
         super.onStart()
@@ -60,11 +69,11 @@ class ImportBookSourceDialog : BaseDialogFragment(), Toolbar.OnMenuItemClickList
 
     override fun onFragmentCreated(view: View, savedInstanceState: Bundle?) {
         binding.toolBar.setTitle(R.string.import_book_source)
+        binding.rotateLoading.show()
         initMenu()
         adapter = SourcesAdapter(requireContext())
         binding.recyclerView.layoutManager = LinearLayoutManager(requireContext())
         binding.recyclerView.adapter = adapter
-        adapter.setItems(viewModel.allSources)
         binding.tvCancel.visible()
         binding.tvCancel.setOnClickListener {
             dismissAllowingStateLoss()
@@ -89,6 +98,31 @@ class ImportBookSourceDialog : BaseDialogFragment(), Toolbar.OnMenuItemClickList
             }
             adapter.notifyDataSetChanged()
             upSelectText()
+        }
+        viewModel.errorLiveData.observe(this, {
+            binding.rotateLoading.hide()
+            errorDialog(it)
+        })
+        viewModel.successLiveData.observe(this, {
+            binding.rotateLoading.hide()
+            if (it > 0) {
+                adapter.setItems(viewModel.allSources)
+            } else {
+                errorDialog(getString(R.string.wrong_format))
+            }
+        })
+        val source = arguments?.getString("source")
+        if (source.isNullOrEmpty()) {
+            dismiss()
+            return
+        }
+        if (source.isAbsUrl()) {
+            viewModel.importSource(source)
+        } else {
+            IntentDataHelp.getData<String>(source)?.let {
+                viewModel.importSource(it)
+                return
+            }
         }
     }
 
@@ -148,9 +182,13 @@ class ImportBookSourceDialog : BaseDialogFragment(), Toolbar.OnMenuItemClickList
         return false
     }
 
-    override fun onDismiss(dialog: DialogInterface) {
-        super.onDismiss(dialog)
-        activity?.finish()
+    private fun errorDialog(msg: String) {
+        alert(getString(R.string.error), msg) {
+            okButton { }
+            onDismiss {
+
+            }
+        }.show()
     }
 
     inner class SourcesAdapter(context: Context) :
