+++ app/src/main/java/io/legado/app/ui/rss/subscription/RuleSubActivity.kt
@@ -13,7 +13,7 @@ import io.legado.app.data.entities.RuleSub
 import io.legado.app.databinding.ActivityRuleSubBinding
 import io.legado.app.databinding.DialogRuleSubEditBinding
 import io.legado.app.lib.dialogs.alert
-import io.legado.app.ui.association.ImportBookSourceActivity
+import io.legado.app.ui.association.ImportBookSourceDialog
 import io.legado.app.ui.association.ImportReplaceRuleActivity
 import io.legado.app.ui.association.ImportRssSourceActivity
 import io.legado.app.ui.widget.recycler.ItemTouchCallback
@@ -74,9 +74,7 @@ class RuleSubActivity : BaseActivity<ActivityRuleSubBinding>(),
     override fun openSubscription(ruleSub: RuleSub) {
         when (ruleSub.type) {
             0 -> {
-                startActivity<ImportBookSourceActivity> {
-                    putExtra("source", ruleSub.url)
-                }
+                ImportBookSourceDialog.start(supportFragmentManager, ruleSub.url)
             }
             1 -> {
                 startActivity<ImportRssSourceActivity> {
